requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
python-dotenv
sentence-transformers
faiss-cpu
pydantic
pydantic-settings
requests
pytest
openai
